Refactor PartitionPruneRule to be reused for MV
